feat: Split out core node package
