+++ plugins/codebuild/backend/src/service/DefaultAwsCodeBuildService.test.ts
@@ -17,7 +17,7 @@ import { ConfigReader } from '@backstage/config';
 import { mockClient } from 'aws-sdk-client-mock';
 import { getVoidLogger } from '@backstage/backend-common';
 import { DefaultAwsCodeBuildService } from './DefaultAwsCodeBuildService';
-import { AwsResourceLocator } from '@aws/aws-core-plugin-for-backstage-common';
+import { AwsResourceLocator } from '@aws/aws-core-plugin-for-backstage-node';
 import {
   DefaultAwsCredentialsManager,
   AwsCredentialProvider,
