+++ plugins/core/common/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aws/aws-core-plugin-for-backstage-common",
-  "description": "Common functionality for the aws-core-plugin-for-backstage plugin",
+  "description": "Common functionality for AWS plugins for Backstage",
   "version": "0.4.2",
   "repository": {
     "type": "git",
@@ -29,22 +29,11 @@
     "prepack": "backstage-cli package prepack",
     "postpack": "backstage-cli package postpack"
   },
-  "devDependencies": {
-    "@backstage/backend-common": "^0.24.0",
-    "@backstage/cli": "^0.27.0",
-    "aws-sdk-client-mock": "^4.0.0"
+  "dependencies": {
+    "@backstage/catalog-model": "^1.6.0",
+    "@backstage/config": "^1.2.0"
   },
   "files": [
     "dist"
-  ],
-  "dependencies": {
-    "@aws-sdk/client-config-service": "^3.511.0",
-    "@aws-sdk/client-resource-explorer-2": "^3.511.0",
-    "@aws-sdk/client-resource-groups-tagging-api": "^3.511.0",
-    "@aws-sdk/types": "^3.511.0",
-    "@backstage/catalog-model": "^1.6.0",
-    "@backstage/config": "^1.2.0",
-    "@backstage/integration-aws-node": "^0.1.12",
-    "winston": "^3.11.0"
-  }
+  ]
 }
