+++ plugins/core/node/README.md
@@ -1,0 +1,3 @@
+# AWS core plugin for Backstage - Node
+
+Functionality to help implement AWS functionality in Backstage backend plugins. For more information see the [documentation](https://github.com/awslabs/backstage-plugins-for-aws).
