+++ plugins/core/node/package.json
@@ -1,25 +1,24 @@
 {
-  "name": "@aws/aws-core-plugin-for-backstage-react",
-  "description": "Web library for the AWS plugins for Backstage",
-  "version": "0.4.1",
+  "name": "@aws/aws-core-plugin-for-backstage-node",
+  "description": "Node library for the AWS plugins for Backstage",
+  "version": "0.0.0",
   "repository": {
     "type": "git",
     "url": "github:awslabs/backstage-plugins-for-aws",
-    "directory": "plugins/core/react"
+    "directory": "plugins/core/node"
   },
   "main": "src/index.ts",
   "types": "src/index.ts",
   "license": "Apache-2.0",
   "publishConfig": {
     "access": "public",
-    "main": "dist/index.esm.js",
+    "main": "dist/index.cjs.js",
     "types": "dist/index.d.ts"
   },
   "backstage": {
-    "role": "web-library",
+    "role": "node-library",
     "pluginId": null
   },
-  "sideEffects": false,
   "scripts": {
     "start": "backstage-cli package start",
     "build": "backstage-cli package build",
@@ -30,21 +29,19 @@
     "postpack": "backstage-cli package postpack"
   },
   "dependencies": {
-    "@backstage/core-components": "^0.14.10",
-    "@backstage/core-plugin-api": "^1.9.3",
-    "@backstage/errors": "^1.2.4",
-    "@material-ui/core": "^4.12.2"
-  },
-  "peerDependencies": {
-    "react": "^16.13.1 || ^17.0.0",
-    "react-dom": "^16.13.1 || ^17.0.0 || ^18.0.0",
-    "react-router-dom": "6.0.0-beta.0 || ^6.3.0"
+    "@aws-sdk/client-config-service": "^3.511.0",
+    "@aws-sdk/client-resource-explorer-2": "^3.511.0",
+    "@aws-sdk/client-resource-groups-tagging-api": "^3.511.0",
+    "@aws-sdk/types": "^3.511.0",
+    "@aws/aws-core-plugin-for-backstage-common": "workspace:^",
+    "@backstage/config": "^1.2.0",
+    "@backstage/integration-aws-node": "^0.1.12",
+    "winston": "^3.11.0"
   },
   "devDependencies": {
+    "@backstage/backend-common": "^0.24.0",
     "@backstage/cli": "^0.27.0",
-    "@backstage/test-utils": "^1.5.10",
-    "@testing-library/jest-dom": "^6.0.0",
-    "@testing-library/react": "^12.1.3"
+    "aws-sdk-client-mock": "^4.0.0"
   },
   "files": [
     "dist"
