+++ plugins/core/node/src/locator/aws-config-locator.test.ts
