+++ plugins/core/node/src/locator/resource-explorer-locator.test.ts
