+++ plugins/core/node/src/locator/resource-locator.ts
