+++ plugins/core/node/src/locator/resource-tagging-api-instance.ts
@@ -12,64 +12,49 @@
  */
 
 import {
-  ResourceExplorer2Client,
-  SearchCommand,
-} from '@aws-sdk/client-resource-explorer-2';
+  GetResourcesCommand,
+  ResourceGroupsTaggingAPIClient,
+} from '@aws-sdk/client-resource-groups-tagging-api';
 import { Logger } from 'winston';
-import {
-  AWS_SDK_CUSTOM_USER_AGENT,
-  AwsResourceLocator,
-} from '@aws/aws-core-plugin-for-backstage-common';
 import { AwsCredentialIdentityProvider } from '@aws-sdk/types';
-import { Config } from '@backstage/config';
 import { DefaultAwsCredentialsManager } from '@backstage/integration-aws-node';
 import { convertResourceTypeString, parseResourceLocatorTags } from './utils';
+import { AWS_SDK_CUSTOM_USER_AGENT } from '@aws/aws-core-plugin-for-backstage-common';
 
-export class AwsResourceExplorerLocator implements AwsResourceLocator {
+export class AwsResourceTaggingApiLocatorInstance {
   public constructor(
     private readonly logger: Logger,
-    private readonly client: ResourceExplorer2Client,
-    private readonly viewArn: string | undefined,
+    private readonly client: ResourceGroupsTaggingAPIClient,
   ) {}
 
-  static async fromConfig(
-    config: Config,
-    options: {
-      logger: Logger;
-    },
-  ) {
-    let region: string | undefined;
-    let accountId: string | undefined;
-    let viewArn: string | undefined;
-
-    const conf = config.getOptionalConfig('aws.locator.resourceExplorer');
-
-    if (conf) {
-      viewArn = conf.getOptionalString('viewArn');
-      accountId = conf.getOptionalString('accountId');
-      region = conf.getOptionalString('region');
-    }
-
-    const credsManager = DefaultAwsCredentialsManager.fromConfig(config);
-
+  static async create(options: {
+    credsManager: DefaultAwsCredentialsManager;
+    accountId: string | undefined;
+    region: string | undefined;
+    logger: Logger;
+  }): Promise<AwsResourceTaggingApiLocatorInstance> {
     let credentialProvider: AwsCredentialIdentityProvider;
+
+    const { accountId, region, credsManager } = options;
 
     if (accountId) {
       credentialProvider = (
-        await credsManager.getCredentialProvider({ accountId })
+        await credsManager.getCredentialProvider({
+          accountId: accountId,
+        })
       ).sdkCredentialProvider;
     } else {
       credentialProvider = (await credsManager.getCredentialProvider())
         .sdkCredentialProvider;
     }
 
-    const client = new ResourceExplorer2Client({
+    const client = new ResourceGroupsTaggingAPIClient({
       region: region,
       customUserAgent: AWS_SDK_CUSTOM_USER_AGENT,
       credentialDefaultProvider: () => credentialProvider,
     });
 
-    return new AwsResourceExplorerLocator(options.logger, client, viewArn);
+    return new AwsResourceTaggingApiLocatorInstance(options.logger, client);
   }
 
   async getResourceArns({
@@ -79,27 +64,24 @@ export class AwsResourceExplorerLocator implements AwsResourceLocator {
     resourceType: string;
     tagString: string;
   }): Promise<string[]> {
-    const tagFilters = parseResourceLocatorTags(tagString);
+    const TagFilters = parseResourceLocatorTags(tagString).map(e => {
+      return {
+        Key: e.key,
+        Values: [e.value],
+      };
+    });
 
     const convertedResourceType = convertResourceTypeString(resourceType);
 
     this.logger.debug(`Retrieving resource ARNs for ${convertedResourceType}`);
 
-    const query = `${tagFilters
-      .map(e => {
-        return `tag:${e.key}=${e.value}`;
-      })
-      .join(' ')} resourcetype:${convertedResourceType}`.trim();
-
-    this.logger.debug(`Resource explorer query: ${query}`);
-
     const response = await this.client.send(
-      new SearchCommand({
-        QueryString: query,
-        ViewArn: this.viewArn,
+      new GetResourcesCommand({
+        ResourceTypeFilters: [convertedResourceType],
+        TagFilters,
       }),
     );
 
-    return response.Resources!.map(e => e.Arn!);
+    return response.ResourceTagMappingList!.map(e => e.ResourceARN!);
   }
 }
