+++ plugins/core/node/src/setupTests.ts
@@ -11,5 +11,4 @@
  * limitations under the License.
  */
 
-export * from './utils';
-export * from './constants';
+export {};
